App fixes across the stack

+++ backend/src/main/java/com/interview/models/Director.java
@@ -7,7 +7,6 @@ import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
-
 import org.hibernate.annotations.CreationTimestamp;
 import org.hibernate.annotations.UpdateTimestamp;
 
@@ -17,7 +16,7 @@ import com.interview.dtos.DirectorDTO;
 public class Director {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id;
+    private Long id;
 
     @Column(name = "first_name")
     private String firstName;
@@ -43,7 +42,7 @@ public class Director {
         this.updatedAt = Instant.now();
     }
 
-    public int getId() {
+    public Long getId() {
         return id;
     }
 
@@ -79,7 +78,7 @@ public class Director {
         this.createdAt = created_at;
     }
 
-    public void setId(int id) {
+    public void setId(Long id) {
         this.id = id;
     }
